src/main/java/com/zytoune/dailygame/controller/advice/ApplicationControllerAdvice.java
package com.zytoune.dailygame.controller.advice;


import com.zytoune.dailygame.exception.NoGameFoundException;
import io.jsonwebtoken.ExpiredJwtException;
import io.jsonwebtoken.JwtException;
import io.jsonwebtoken.MalformedJwtException;
import lombok.extern.slf4j.Slf4j;
import org.springframework.http.HttpStatus;
import org.springframework.http.ProblemDetail;
import org.springframework.security.authentication.BadCredentialsException;
import org.springframework.security.authorization.AuthorizationDeniedException;
import org.springframework.web.bind.annotation.ExceptionHandler;
import org.springframework.web.bind.annotation.ResponseBody;
import org.springframework.web.bind.annotation.ResponseStatus;
import org.springframework.web.bind.annotation.RestControllerAdvice;

import java.nio.file.AccessDeniedException;
import java.security.SignatureException;
import java.util.Map;

@Slf4j
@RestControllerAdvice
public class ApplicationControllerAdvice {

    @ResponseStatus(HttpStatus.UNAUTHORIZED)
    @ExceptionHandler(value = Exception.class)
    public ProblemDetail exceptionsHandler(Exception exception){
        log.error(exception.getMessage(), exception);
        ProblemDetail problemDetail = ProblemDetail.forStatusAndDetail(HttpStatus.UNAUTHORIZED, "Accès refusé");
        problemDetail.setProperty("erreur", exception.getMessage());
        return problemDetail;
    }

    @ResponseStatus(HttpStatus.FORBIDDEN)
    @ExceptionHandler(value = {AuthorizationDeniedException.class, AccessDeniedException.class})
    public @ResponseBody ProblemDetail accessDeniedException(Exception exception){
        log.error(exception.getMessage(), exception);
        ProblemDetail problemDetail = ProblemDetail.forStatusAndDetail(HttpStatus.FORBIDDEN, "Accès refusé");
        problemDetail.setProperty("erreur", exception.getMessage());
        return problemDetail;
    }

    @ResponseStatus(HttpStatus.UNAUTHORIZED)
    @ExceptionHandler(value = RuntimeException.class)
    public @ResponseBody ProblemDetail runtimeException(Exception exception){
        log.error(exception.getMessage(), exception);
        ProblemDetail problemDetail = ProblemDetail.forStatusAndDetail(HttpStatus.INTERNAL_SERVER_ERROR, "Erreur interne : " + exception.getMessage());
        problemDetail.setProperty("erreur", exception.getMessage());
        return problemDetail;
    }

    @ResponseStatus(HttpStatus.NOT_FOUND)
    @ExceptionHandler(value = NoGameFoundException.class)
    public @ResponseBody ProblemDetail noGameFoundException(Exception exception){
        log.error(exception.getMessage(), exception);
        ProblemDetail problemDetail = ProblemDetail.forStatusAndDetail(HttpStatus.NOT_FOUND, "Le jeu demandé n'a pas été trouvé en base de données ou n'existe pas");
        problemDetail.setProperty("erreur", exception.getMessage());
        return problemDetail;
    }

}
